requirements.txt
aiogram==3.9.0
python-dotenv
flask
aiohttp==3.7.4